% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{base85_decode}
\alias{base85_decode}
\title{Z85 Decoding}
\usage{
base85_decode(encoded_string)
}
\arguments{
\item{encoded_string}{A string}
}
\description{
Decodes a Z85 encoded string back to binary
}
