% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{base91_decode}
\alias{base91_decode}
\title{basE91 Decoding}
\usage{
base91_decode(encoded_string)
}
\arguments{
\item{encoded_string}{A string}
}
\description{
Decodes a basE91 encoded string back to binary
}
