% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_encoding.r
\name{base91_encode}
\alias{base91_encode}
\title{basE91 Encoding}
\usage{
base91_encode(rawdata, quote_character = "\"")
}
\arguments{
\item{rawdata}{A raw vector}

\item{quote_character}{The character to use in the encoding, replacing the double quote character. Must be a single quote "'", double quote "\"" or dash "-".}
}
\value{
A string representation of the raw vector.
}
\description{
Encodes binary data (a raw vector) as ascii text using basE91 encoding format
}
\details{
basE91 (capital E for stylization) is a binary to ascii encoding format created by Joachim Henke in 2005.
The overhead (extra bytes used relative to binary) is 22.97\% on average. In comparison, base 64 encoding has an overhead of 33.33\%.
The original encoding uses a dictionary of 91 out of 94 printable ASCII characters excluding - (dash), \ (backslash) and ' (single quote).
The original encoding does include double quote characters, which are less than ideal for strings in R. Therefore,
you can use the `quote_character` parameter to substitute dash or single quote.
}
\references{
http://base91.sourceforge.net/
}
