% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{blosc_shuffle_raw}
\alias{blosc_shuffle_raw}
\title{Shuffle a raw vector}
\usage{
blosc_shuffle_raw(x, bytesofsize)
}
\arguments{
\item{x}{A raw vector.}

\item{bytesofsize}{Either \code{4} or \code{8}.}
}
\value{
The shuffled vector
}
\description{
Shuffles a raw vector using BLOSC shuffle routines.
}
\examples{
x <- serialize(1L:1000L, NULL)
xshuf <- blosc_shuffle_raw(x, 4)
xunshuf <- blosc_unshuffle_raw(xshuf, 4)
}
