% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{blosc_unshuffle_raw}
\alias{blosc_unshuffle_raw}
\title{Un-shuffle a raw vector}
\usage{
blosc_unshuffle_raw(x, bytesofsize)
}
\arguments{
\item{x}{The raw vector}

\item{bytesofsize}{Either 4 or 8}
}
\value{
The unshuffled vector
}
\description{
A function for un-shuffling a raw vector using BLOSC un-shuffle routines
}
\examples{
x <- serialize(1L:1000L, NULL)
xshuf <- blosc_shuffle_raw(x, 4)
xunshuf <- blosc_unshuffle_raw(xshuf, 4)
}
