% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{catquo}
\alias{catquo}
\title{catquo}
\usage{
catquo(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[=cat]{cat()}}.}
}
\description{
Prints a string with single quotes on a new line.
}
