% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeCasteljau.R
\name{DeCasteljau}
\alias{DeCasteljau}
\title{Title}
\usage{
DeCasteljau(segments, keyTimes = NULL, times)
}
\arguments{
\item{segments}{x}

\item{keyTimes}{x}

\item{times}{x}
}
\value{
x
}
\description{
Title
}
\examples{
2+2
}
