% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_encoding.r
\name{decode_source}
\alias{decode_source}
\title{Decode a compressed string}
\usage{
decode_source(string)
}
\arguments{
\item{string}{The string to decode}
}
\value{
The original (decoded) string.
}
\description{
A helper function for encoding and compressing a file or string to ascii using `base91_encode` and `qserialize` with the highest compression level.
}
