% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_encoding.r
\name{encode_source}
\alias{encode_source}
\title{Encode and compress a file or string}
\usage{
encode_source(file = NULL, string = NULL, width = 120)
}
\arguments{
\item{file}{The file to encode (if `string` is not NULL)}

\item{string}{The string to encode (if `file` is not NULL)}

\item{width}{The output will be broken up into individual strings, with `width` being the longest allowable string.}
}
\value{
A CharacterVector in base91 representing the compressed original file or string.
}
\description{
A helper function for encoding and compressing a file or string to ascii using `base91_encode` and `qserialize` with the highest compression level.
}
