% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{get_altrep_class_info}
\alias{get_altrep_class_info}
\title{Get the class information of an ALTREP object}
\usage{
get_altrep_class_info(obj)
}
\arguments{
\item{obj}{The ALTREP class name}
}
\value{
The class information (class name and package name) of an ALTREP object, a character vector of length two.
If the object is not an ALTREP object, returns NULL.
}
\description{
Gets the formal name of the class and package of an ALTREP object
}
\examples{
get_altrep_class_info(1:5)
}
