% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{lz4_decompress_raw}
\alias{lz4_decompress_raw}
\title{lz4 decompression}
\usage{
lz4_decompress_raw(x)
}
\arguments{
\item{x}{A Raw Vector}
}
\value{
The uncompressed data
}
\description{
Decompresses of raw vector
}
\examples{
x <- 1:1e6
xserialized <- serialize(x, connection=NULL)
xcompressed <- lz4_compress_raw(xserialized, compress_level = 1)
xrecovered <- unserialize(lz4_decompress_raw(xcompressed))
}
