% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{qdeserialize}
\alias{qdeserialize}
\title{qdeserialize}
\usage{
qdeserialize(x, use_alt_rep=FALSE, strict=FALSE)
}
\arguments{
\item{x}{A raw vector.}

\item{use_alt_rep}{Use ALTREP when reading in string data (default \code{FALSE}). On R versions prior to 3.5.0, this parameter does nothing.}

\item{strict}{Whether to throw an error or just report a warning (default: \code{FALSE}, i.e. report warning).}
}
\value{
The de-serialized object.
}
\description{
Reads an object from a raw vector.
}
\details{
See \code{\link[=qserialize]{qserialize()}} for additional details and examples.
}
