% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{qdump}
\alias{qdump}
\title{qdump}
\arguments{
\item{file}{the file name/path.}
}
\value{
The uncompressed serialization
}
\description{
Exports uncompressed serialization to a list of Raw Vectors.  For testing purposes.
}
