% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{qread_pipe}
\alias{qread_pipe}
\title{qread_pipe}
\usage{
qread_pipe(scon, use_alt_rep=FALSE, strict=FALSE, check_mode = TRUE)
}
\arguments{
\item{scon}{A string or an R connection (opened in "rb" mode).}

\item{use_alt_rep}{Use alt rep when reading in string data.  Default: FALSE.  (Note: on R versions earlier than 3.5.0, this parameter does nothing.)}

\item{strict}{Whether to throw an error or just report a warning (Default: FALSE, report warning)}

\item{check_mode}{Check mode of connection object (should be "rb"), default: TRUE}
}
\value{
The de-serialized object
}
\description{
Reads an object from a pipe
}
\details{
See `?qsave_pipe` for additional details and examples.
}
