% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{qread_ptr}
\alias{qread_ptr}
\title{qread_ptr}
\usage{
qread_ptr(pointer, length, use_alt_rep=FALSE, strict=FALSE)
}
\arguments{
\item{pointer}{An external pointer to memory.}

\item{length}{The length of the object in memory.}

\item{use_alt_rep}{Use ALTREP when reading in string data (default \code{FALSE}). On R versions prior to 3.5.0, this parameter does nothing.}

\item{strict}{Whether to throw an error or just report a warning (default: \code{FALSE}, i.e. report warning).}
}
\value{
The de-serialized object.
}
\description{
Reads an object from an external pointer.
}
