% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qread_url.R
\name{qread_url}
\alias{qread_url}
\title{qread_url}
\usage{
qread_url(url, buffer_size, ...)
}
\arguments{
\item{url}{The URL where the object is stored}

\item{buffer_size}{The buffer size used to read in data (default \code{16777216L} i.e. 16 MB)}

\item{...}{Arguments passed to \code{\link[=qdeserialize]{qdeserialize()}}}
}
\value{
The de-serialized object.
}
\description{
A helper function that reads data from the internet to memory and deserializes the object with \code{\link[=qdeserialize]{qdeserialize()}}.
}
\details{
See \code{\link[=qdeserialize]{qdeserialize()}} for additional details.
}
\examples{
\dontrun{
x <- qread_url("http://example_url.com/my_file.qs")
}
}
