% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs-package.R
\docType{package}
\name{qs-package}
\alias{qs}
\alias{qs-package}
\title{qs: Quick Serialization of R Objects}
\description{
Provides functions for quickly writing and reading any R object to and from disk.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/traversc/qs}
  \item Report bugs at \url{https://github.com/traversc/qs/issues}
}

}
\author{
\strong{Maintainer}: Travers Ching \email{traversc@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Yann Collet (Yann Collet is the author of the bundled zstd, lz4 and xxHash code) [contributor, copyright holder]
  \item Facebook, Inc. (Facebook is the copyright holder of the bundled zstd code) [copyright holder]
  \item Reichardt Tino (Contributor/copyright holder of zstd bundled code) [contributor, copyright holder]
  \item Skibinski Przemyslaw (Contributor/copyright holder of zstd bundled code) [contributor, copyright holder]
  \item Mori Yuta (Contributor/copyright holder of zstd bundled code) [contributor, copyright holder]
  \item Romain Francois (Derived example/tutorials for ALTREP structures) [contributor, copyright holder]
  \item Francesc Alted (Shuffling routines derived from Blosc library) [contributor, copyright holder]
  \item Bryce Chamberlain \email{superchordate@gmail.com} (qsavem and qload functions) [contributor]
  \item Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID}) (documentation) [contributor]
}

}
\keyword{internal}
