% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{qs_set_blocksize}
\alias{qs_set_blocksize}
\title{Set Blocksize}
\arguments{
\item{blocksize}{The blocksize in bytes}
}
\description{
Changes the compression blocksize.  Default: 2^19 (512 Kb).  Mostly for testing purposes, don't change otherwise.  
If you change the blocksize, you'll need to set it every time you load an object.
}
