% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{qs_show_warnings}
\alias{qs_show_warnings}
\title{Show Warnings}
\arguments{
\item{show_warnings}{A boolean showing a warning when attempting to serialize an unsupported object type.  Default: TRUE.}
}
\description{
Show warnings during compression.
}
