% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{qs_use_alt_rep}
\alias{qs_use_alt_rep}
\title{Use alt-rep}
\arguments{
\item{s}{A boolean to determine whether `qread` uses alt-rep.  Default: TRUE.}
}
\description{
Changes whether qread uses the alt-rep system
}
