% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{qsave}
\alias{qsave}
\title{qsave}
\arguments{
\item{file}{the file name/path.}

\item{compress_level}{The compression level (-50 to 22).  Default -1.  Higher values tend to have a better compression ratio, while lower values/negative values tend to be quicker.  Values with good compression/speed tradeoffs seem to be -1, 1 and 4.}
}
\value{
The de-serialized object
}
\description{
Saves (serializes) an object to disk.
}
\examples{
x1 <- data.frame(int = sample(1e5, replace=TRUE), 
                 num = rnorm(1e5), 
                 char = randomStrings(1e5), stringsAsFactors = FALSE)
qsave(x1, "/tmp/mydata.qs")
}
