% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{randomStrings}
\alias{randomStrings}
\title{Generate random strings}
\usage{
randomStrings(N, string_size)
}
\arguments{
\item{N}{The number of random strings to generate}

\item{string_size}{The number of characters in each string (default 50).}
}
\value{
A character vector of random alpha-numeric strings.
}
\description{
A function for generating a character vector of random strings, for testing purposes.
}
\examples{
randomStrings(N=10, string_size=20) # returns 10 alphanumeric strings of length 20
randomStrings(N=100, string_size=200) # returns 100 alphanumeric strings of length 200
}
