% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{set_trust_promises}
\alias{set_trust_promises}
\title{Allow for serialization/deserialization of promises}
\usage{
set_trust_promises(value)
}
\arguments{
\item{value}{a boolean \code{TRUE} or \code{FALSE}}
}
\value{
The previous value of the global variable \code{trust_promises}
}
\description{
Allow for serialization/deserialization of promises
}
\examples{
set_trust_promises(TRUE)
}
