% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\docType{data}
\name{starnames}
\alias{starnames}
\title{Official list of IAU Star Names}
\format{A `data.frame` with official IAU star names and several properties, such as coordinates.}
\source{
\href{https://www.iau.org/public/themes/naming_stars/}{Naming Stars | International Astronomical Union.}
}
\usage{
data(starnames)
}
\description{
Data from the International Astronomical Union.  
An official list of the 336 internationally recognized named stars,
updated as of June 1, 2018.
}
\examples{
data(starnames)
}
\references{
E Mamajek et. al. (2018), 
\emph{WG Triennial Report (2015-2018) - Star Names}, Reports on Astronomy, 22 Mar 2018.
}
\keyword{datasets}
