% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{unregister_altrep_class}
\alias{unregister_altrep_class}
\title{Unegister ALTREP class for serialization}
\usage{
unregister_altrep_class(classname, pkgname)
}
\arguments{
\item{classname}{The ALTREP class name}

\item{pkgname}{The package the ALTREP class comes from}
}
\description{
Unegister an ALTREP class to not use base R serialization.
}
\examples{
unregister_altrep_class("compact_intseq", "base")
}
