% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{zstd_compress_raw}
\alias{zstd_compress_raw}
\title{Zstd compression}
\arguments{
\item{x}{A Raw Vector}

\item{compress_level}{The compression level (-50 to 22)}
}
\value{
The compressed data
}
\description{
Compression of raw vector
}
