% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{convertToAlt}
\alias{convertToAlt}
\title{Convert character vector to alt-rep}
\usage{
convertToAlt(x)
}
\arguments{
\item{x}{The character vector}
}
\value{
The character vector in alt-rep form
}
\description{
A function for generating a alt-rep object from a character vector, for testing purposes.
}
