% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{convertToAlt}
\alias{convertToAlt}
\title{Convert character vector to alt-rep}
\usage{
convertToAlt(x)
}
\arguments{
\item{x}{The character vector}
}
\value{
The character vector in alt-rep form
}
\description{
A function for generating a alt-rep object from a character vector, for users to experiment with the alt-rep system.
This function is not available in R versions earlier than 3.5.0.
}
\examples{
xalt <- convertToAlt(randomStrings(N=10, string_size=20))
xalt2 <- convertToAlt(c("a", "b", "c"))
}
