% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_encoding.R
\name{decode_source}
\alias{decode_source}
\title{Decode a compressed string}
\usage{
decode_source(string)
}
\arguments{
\item{string}{A string to decode.}
}
\value{
The original (decoded) object.
}
\description{
A helper function for encoding and compressing a file or string to ASCII using \code{\link[=base91_encode]{base91_encode()}} and \code{\link[=qserialize]{qserialize()}} with the highest compression level.
}
\seealso{
\code{\link[=encode_source]{encode_source()}} for more details.
}
