% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_encoding.R
\name{encode_source}
\alias{encode_source}
\title{Encode and compress a file or string}
\usage{
encode_source(file = NULL, string = NULL, width = 120)
}
\arguments{
\item{file}{The file to encode (if \code{string} is not NULL)}

\item{string}{The string to encode (if \code{file} is not NULL)}

\item{width}{The output will be broken up into individual strings, with \code{width} being the longest allowable string.}
}
\value{
A character vector in base91 representing the compressed original file or string.
}
\description{
A helper function for encoding and compressing a file or string to ASCII using \code{\link[=base91_encode]{base91_encode()}} and \code{\link[=qserialize]{qserialize()}} with the highest compression level.
}
