% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{lz4_decompress_raw}
\alias{lz4_decompress_raw}
\title{lz4 decompression}
\usage{
lz4_decompress_raw(x)
}
\arguments{
\item{x}{A raw vector.}
}
\value{
The de-serialized object.
}
\description{
Decompresses an lz4 compressed raw vector.
}
\examples{
x <- 1:1e6
xserialized <- serialize(x, connection=NULL)
xcompressed <- lz4_compress_raw(xserialized, compress_level = 1)
xrecovered <- unserialize(lz4_decompress_raw(xcompressed))
}
