% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{qdump}
\alias{qdump}
\title{qdump}
\usage{
qdump(file)
}
\arguments{
\item{file}{A file name/path.}
}
\value{
The uncompressed serialization.
}
\description{
Exports the uncompressed binary serialization to a list of raw vectors. For testing purposes and exploratory purposes mainly.
}
\examples{
x <- data.frame(int = sample(1e3, replace=TRUE),
        num = rnorm(1e3),
        char = sample(starnames$`IAU Name`, 1e3, replace=TRUE),
        stringsAsFactors = FALSE)
myfile <- tempfile()
qsave(x, myfile)
x2 <- qdump(myfile)
}
