% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{qread}
\alias{qread}
\title{qread}
\arguments{
\item{file}{the file name/path}
}
\value{
The de-serialized object
}
\description{
Reads a object in a file serialized to disk
}
\examples{
x1 <- data.frame(int = sample(1e5, replace=TRUE), 
                 num = rnorm(1e5), 
                 char = randomStrings(1e5), stringsAsFactors = FALSE)
qsave(x1, "/tmp/mydata.qs")
x2 <- qread("/tmp/mydata.qs")
identical(x1, x2) # returns true
}
