% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{qread_fd}
\alias{qread_fd}
\title{qread_fd}
\usage{
qread_fd(fd, use_alt_rep=FALSE, strict=FALSE)
}
\arguments{
\item{fd}{A file descriptor}

\item{use_alt_rep}{Use alt rep when reading in string data. Default: FALSE. (Note: on R versions earlier than 3.5.0, this parameter does nothing.)}

\item{strict}{Whether to throw an error or just report a warning (Default: FALSE, report warning)}
}
\value{
The de-serialized object
}
\description{
Reads an object from a file descriptor
}
\details{
See `?qsave_fd` for additional details and examples.
}
