% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{qread_handle}
\alias{qread_handle}
\title{qread_handle}
\usage{
qread_handle(handle, use_alt_rep=FALSE, strict=FALSE)
}
\arguments{
\item{handle}{A windows handle external pointer}

\item{use_alt_rep}{Use alt rep when reading in string data. Default: FALSE. (Note: on R versions earlier than 3.5.0, this parameter does nothing.)}

\item{strict}{Whether to throw an error or just report a warning (Default: FALSE, report warning)}
}
\value{
The de-serialized object
}
\description{
Reads an object from a windows handle
}
\details{
See `?qsave_handle` for additional details and examples.
}
