% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{qread_ptr}
\alias{qread_ptr}
\title{qread_ptr}
\usage{
qread_ptr(pointer, length, use_alt_rep=FALSE, strict=FALSE)
}
\arguments{
\item{pointer}{An external pointer to memory}

\item{length}{the length of the object in memory}

\item{use_alt_rep}{Use alt rep when reading in string data.  Default: FALSE.  (Note: on R versions earlier than 3.5.0, this parameter does nothing.)}

\item{strict}{Whether to throw an error or just report a warning (Default: FALSE, report warning)}
}
\value{
The de-serialized object
}
\description{
Reads an object from a external pointer
}
