% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{randomStrings}
\alias{randomStrings}
\title{Generate random strings}
\usage{
randomStrings(N, string_size)
}
\arguments{
\item{N}{The number of random strings to generate}

\item{string_size}{The number of characters in each string (default 50).}
}
\value{
A character vector of random alpha-numeric strings.
}
\description{
A function for generating a character vector of random strings, for testing purposes.
}
