% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{zstd_compressBound}
\alias{zstd_compressBound}
\title{Zstd CompressBound}
\usage{
zstd_compressBound(size)
}
\arguments{
\item{size}{An integer size}
}
\value{
maximum compressed size
}
\description{
Returns the maximum compressed size of an object of length `size`
}
