% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{zstd_decompress_raw}
\alias{zstd_decompress_raw}
\title{Zstd decompression}
\usage{
zstd_decompress_raw(x)
}
\arguments{
\item{x}{A Raw Vector}
}
\value{
The uncompressed data
}
\description{
Decompresses of raw vector
}
\examples{
x <- 1:1e6
xserialized <- serialize(x, connection=NULL)
xcompressed <- zstd_compress_raw(xserialized, compress_level = 1)
xrecovered <- unserialize(zstd_decompress_raw(xcompressed))
}
