% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{zstd_decompress_raw}
\alias{zstd_decompress_raw}
\title{Zstd decompression}
\usage{
zstd_decompress_raw(x)
}
\arguments{
\item{x}{A raw vector.}
}
\value{
The de-serialized object.
}
\description{
Decompresses a zstd compressed raw vector.
}
\examples{
x <- 1:1e6
xserialized <- serialize(x, connection=NULL)
xcompressed <- zstd_compress_raw(xserialized, compress_level = 1)
xrecovered <- unserialize(zstd_decompress_raw(xcompressed))
}
